Release Soccer Stars Assist v1.0.0 Android APK
Complete overlay assist app with screen capture, aim ruler, ball trajectory
prediction, settings UI, physics.json calibration, performance throttling,
unit tests, and release APK build.

+++ soccer-stars-android/app/src/main/java/com/accessibility/soccerstars/vision/GameDetector.kt
@@ -211,7 +211,10 @@ class GameDetector(
 
             val start = near.minByOrNull { hypot(it.first - puck.x, it.second - puck.y) } ?: continue
             val end = near.maxByOrNull {
-                hypot(it.first - start.first, it.second - start.second)
+                hypot(
+                    (it.first - start.first).toDouble(),
+                    (it.second - start.second).toDouble(),
+                )
             } ?: continue
 
             val length = hypot(
